% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{block_index}
\alias{block_index}
\title{Create block index}
\usage{
block_index(list_labels)
}
\arguments{
\item{list_labels}{Position labels that define the blocks.}
}
\value{
Block index of each position.
}
\description{
Get the index of each block in a list.
}
\examples{
list_labels <- list(2, 2, 3, 3, 3, 1, 1)
block_index(list_labels)
}
