% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{boot_ci}
\alias{boot_ci}
\title{Bootstrap CI}
\usage{
boot_ci(stat, ci = 95, R = 1000)
}
\arguments{
\item{stat}{Vector of observations for some statistic.}

\item{ci}{Confidence interval in percentile units.}

\item{R}{Number of bootstrap iterations to use when calculating statistics.}
}
\value{
Results with \code{mean}, \code{lower}, and \code{upper} columns indicating the mean
statistic and the lower and upper bounds of the CI based on a bootstrap
procedure.
}
\description{
Mean and bootstrap confidence interval.
}
\examples{
# Calculate analysis results
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
results <- spc(data)

# Calculate statistics grouped by input (serial) position
library(magrittr)
library(dplyr)
results \%>\% group_by(input) \%>\% summarise(boot_ci(recall))
}
