% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{category_clustering}
\alias{category_clustering}
\title{Category clustering}
\usage{
category_clustering(data, category_key)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{category_key}{Name of column with category labels.}
}
\value{
Results with \code{subject}, lbc\verb{, and }arc\verb{columns. The}lbc`` column
indicates the list-based clustering measure. The \code{arc} column contains the
adjusted ratio of clustering measure.
}
\description{
Measures of category clustering of recall sequences.
}
\examples{
# Get lists with a mixture of categories
raw <- sample_data("Morton2013")
mixed <- raw[raw$list_type == "mixed", ]

# Calculate category clustering measures
data <- merge_free_recall(mixed, list_keys = list("category"))
stats <- category_clustering(data, "category")
head(stats)
}
