% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{category_crp}
\alias{category_crp}
\title{Category conditional response probability}
\usage{
category_crp(
  data,
  category_key,
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{category_key}{Name of column with category labels.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject}, \code{prob}, \code{actual}, and \code{possible} columns.
The \code{prob} column indicates conditional response probability. The \code{actual}
column indicates the count of transitions actually made at a given lag. The
\code{possible} column indicates the number of transitions that could have been
made, given item availability (previously recalled items are excluded).
}
\description{
Response probability of within-category transitions, conditional on
availability.
}
\examples{
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw, study_keys = list("category"))
cat_crp <- category_crp(data, "category")
head(cat_crp)
}
