% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{check_data}
\alias{check_data}
\title{Check raw free recall data}
\usage{
check_data(data)
}
\arguments{
\item{data}{Raw free recall data in standard format.}
}
\description{
Run checks on raw (unmerged) free recall data.
}
\examples{
# Create data with a required column missing
raw <- data.frame(
  subject = list(1, 1),
  list = list(1, 1),
  position = list(1, 2),
  item = list("a", "b")
)

# Checking this dataset will display an error
# check_data(raw)
}
