% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{distance_crp}
\alias{distance_crp}
\title{Distance conditional response probability}
\usage{
distance_crp(
  data,
  index_key,
  distances,
  edges,
  centers = NULL,
  count_unique = FALSE,
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{index_key}{Name of column containing the index of each item in the
distances matrix.}

\item{distances}{Items x items matrix of pairwise distances.}

\item{edges}{Edges of bins to apply to the distances.}

\item{centers}{Centers to label each bin. If not specified, the center point
between edges will be used.}

\item{count_unique}{If TRUE, possible transitions to a given distance bin
will only count once for a given transition.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject}, \code{bin}, \code{prob}, \code{actual}, and \code{possible}
columns. The \code{prob} column indicates conditional response probability. The
\code{actual} column indicates the count of transitions actually made at a given
distance bin. The \code{possible} column indicates the number of transitions
that could have been made, given item availability (previously recalled
items are excluded).
}
\description{
Response probability as a function of a distance measure, conditional on
item availability.
}
\examples{
# Load data and item-item distances
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
d <- sample_distances("Morton2013")

# Prepare bin definitions
percentiles <- pracma::linspace(.01, .99, 10)
edges <- quantile(pracma::squareform(d$distances), percentiles)

# Calculate distance CRP
data$item_index <- pool_index(data$item, d$items)
crp <- distance_crp(data, "item_index", d$distances, edges)
}
