% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{distance_rank}
\alias{distance_rank}
\title{Distance rank}
\usage{
distance_rank(
  data,
  index_key,
  distances,
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{index_key}{Name of column containing the index of each item in the
distances matrix.}

\item{distances}{Items x items matrix of pairwise distances.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject} and \code{rank} columns. A \code{rank} of 1 indicates
that the smallest distance transition was always made, while 0.5 indicates
chance clustering, and 0 indicates that the largest distance transition was
always made.
}
\description{
Rank of transition distances.
}
\examples{
# Load data and item-item distances
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
d <- sample_distances("Morton2013")

# Calculate distance rank
data$item_index <- pool_index(data$item, d$items)
ranks <- distance_rank(data, "item_index", d$distances)
}
