% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{distance_rank_shifted}
\alias{distance_rank_shifted}
\title{Distance rank shifted}
\usage{
distance_rank_shifted(
  data,
  index_key,
  distances,
  max_shift,
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{index_key}{Name of column containing the index of each item in the
distances matrix.}

\item{distances}{Items x items matrix of pairwise distances.}

\item{max_shift}{Maximum number of items back for which to rank distances.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject}, \code{shift}, and \code{rank} columns. A \code{rank} of 1
indicates that the smallest distance transition was always made, while 0.5
indicates chance clustering, and 0 indicates that the largest distance
transition was always made.
}
\description{
Rank of transition distances, shifted by recall lag.
}
\examples{
# Load data and item-item distances
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
d <- sample_distances("Morton2013")

# Calculate distance rank
data$item_index <- pool_index(data$item, d$items)
ranks <- distance_rank_shifted(data, "item_index", d$distances, max_shift = 3)
head(ranks)
}
