% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{filter_data}
\alias{filter_data}
\title{Filter free recall data}
\usage{
filter_data(
  data,
  subjects = NULL,
  lists = NULL,
  trial_type = NULL,
  positions = NULL,
  inputs = NULL,
  outputs = NULL
)
}
\arguments{
\item{data}{Raw or merged data to filter.}

\item{subjects}{Subject or subjects to include.}

\item{lists}{List or lists to include.}

\item{trial_type}{Trial type to include.}

\item{positions}{Position or positions to include.}

\item{inputs}{Input position or positions to include.}

\item{outputs}{Output position or positions to include.}
}
\value{
The filtered subset of data.
}
\description{
Filter raw or scored data to get a subset of trials or study/recall pairings.
}
\examples{
# Filter data to get study events for subject 1
subjects_list <- list(1, 1, 2, 2)
study_lists <- list(
  list("a", "b"), list("c", "d"), list("e", "f"), list("g", "h")
)
recall_lists <- list(list("b"), list("d", "c"), list("f", "e"), list())
raw <- table_from_lists(subjects_list, study_lists, recall_lists)
filter_data(raw, subjects = 1, trial_type = "study")

# Filtered scored data to get subject 2
data <- merge_free_recall(raw)
filter_data(data, subjects = 2)
}
