% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{group_stats}
\alias{group_stats}
\title{Group statistics}
\usage{
group_stats(results, stat, ..., ci = 95, R = 1000)
}
\arguments{
\item{results}{Data frame with results of an analysis.}

\item{stat}{Column to use when calculating statistics.}

\item{...}{Columns to group by.}

\item{ci}{Confidence interval in percentile units.}

\item{R}{Number of bootstrap iterations to use when calculating statistics.}
}
\value{
Results with additional \code{mean}, \code{lower}, and \code{upper} columns
indicating the mean statistic and the lower and upper bounds of the CI
based on a bootstrap procedure.
}
\description{
Mean and confidence interval for each group.
}
\examples{
# Calculate analysis results
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
results <- spc(data)

# Calculate statistics grouped by input (serial) position
group_stats(results, recall, input)
}
