% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{lag_crp}
\alias{lag_crp}
\title{Lag conditional response probability}
\usage{
lag_crp(
  data,
  lag_key = "input",
  count_unique = FALSE,
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{lag_key}{Name of column to use when calculating lag between recalled
items.}

\item{count_unique}{If TRUE, possible transitions of the same lag will only
be incremented once per transition.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject}, \code{lag}, \code{prob}, \code{actual}, and \code{possible}
columns. The \code{prob} column indicates conditional response probability. The
\code{actual} column indicates the count of transitions actually made at a given
lag. The \code{possible} column indicates the number of transitions that could
have been made, given item availability (previously recalled items are
excluded).
}
\description{
Probability of recalling an item as a function of its lag from the previous
recall, conditional on it being available for recall.
}
\examples{
# All transitions included
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw, study_keys = list("category"))
head(lag_crp(data))

# Excluding the first three output positions (need to include non-recalled
# items specifically so they aren't excluded as possible items to recall)
head(lag_crp(data, item_query = "output > 3 or not recall"))

# Including within-category transitions only
head(lag_crp(data, test_key = "category", test = function(x, y) x == y))
}
