% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{lag_crp_compound}
\alias{lag_crp_compound}
\title{Compound lag conditional response probability}
\usage{
lag_crp_compound(
  data,
  lag_key = "input",
  count_unique = FALSE,
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{lag_key}{Name of column to use when calculating lag between recalled
items.}

\item{count_unique}{If TRUE, possible transitions of the same lag will only
be incremented once per transition.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject}, \code{previous}, \code{current}, \code{prob}, \code{actual}, and
\code{possible} columns. The \code{prob} column indicates conditional response
probability. The \code{actual} column indicates the count of transitions
actually made at a given combination of \code{previous} and \code{current} lags. The
\code{possible} column indicates the number of transitions that could have been
made, given item availability (previously recalled items are excluded).
}
\description{
Response probability as a function of the lag of current and prior
transitions, conditional on item availability.
}
\examples{
# Create short example list with three recalls
subjects <- list(1)
study <- list(list("absence", "hollow", "pupil", "fountain"))
recall <- list(list("fountain", "hollow", "absence"))
raw <- table_from_lists(subjects, study, recall)
data <- merge_free_recall(raw)

# Display compound CRP for previous lags of -3 and -2
head(lag_crp_compound(data), 14)
}
