% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{lag_rank}
\alias{lag_rank}
\title{Lag rank}
\usage{
lag_rank(
  data,
  lag_key = "input",
  item_query = NULL,
  test_key = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Merged study and recall data.}

\item{lag_key}{Name of column to use when calculating lag between recalled
items.}

\item{item_query}{Query string to select items to include in the pool of
possible recalls to be examined.}

\item{test_key}{Name of column with labels to use when testing transitions
for inclusion.}

\item{test}{Function that takes in previous and current item values and
returns TRUE for transitions that should be included.}
}
\value{
Results with \code{subject} and \code{rank} columns. The \code{rank} indicates how
strongly recalls were clustered by input position lag relative to the
possible items that could have been recalled on each transition. A \code{rank}
of 1 indicates that the lowest lag item was always recalled. A \code{rank} of
0.5 indicates chance clustering. A \code{rank} of 0 indicates that the highest
lag item was always recalled.
}
\description{
Calculate the mean rank of absolute input position lags between adjacent
recalls, relative to possible absolute lags.
}
\examples{
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw, study_keys = list("category"))
head(lag_rank(data))
}
