% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{merge_free_recall}
\alias{merge_free_recall}
\title{Score standard free recall data}
\usage{
merge_free_recall(
  data,
  merge_keys = NULL,
  list_keys = NULL,
  study_keys = NULL,
  recall_keys = NULL,
  position_key = "position"
)
}
\arguments{
\item{data}{Free recall data in Psifr format. Must have subject, list,
trial_type, position, and item columns.}

\item{merge_keys}{Columns to use to designate events to merge. Default is
list(‘subject’, ‘list’, ‘item’), which will merge events related to the same
item, but only within list.}

\item{list_keys}{Columns that apply to both study and recall events.}

\item{study_keys}{Columns that only apply to study events.}

\item{recall_keys}{Columns that only apply to recall events.}

\item{position_key}{Column indicating the position of each item in either
the study list or the recall sequence.}
}
\value{
Merged information about study and recall events. Each row
corresponds to one unique input/output pair.
}
\description{
Merge study and recall events that have the same subject, list, and item.
}
\examples{
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
}
