% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{merge_lists}
\alias{merge_lists}
\title{Score study and recall events}
\usage{
merge_lists(
  study,
  recall,
  merge_keys = NULL,
  list_keys = NULL,
  study_keys = NULL,
  recall_keys = NULL,
  position_key = "position"
)
}
\arguments{
\item{study}{Information about all study events.
Should have one row for each study event.}

\item{recall}{Information about all recall events.
Should have one row for each recall attempt.}

\item{merge_keys}{Columns to use to designate events to merge.
Default is list(‘subject’, ‘list’, ‘item’), which will merge events
related to the same item, but only within list.}

\item{list_keys}{Columns that apply to both study and recall events.}

\item{study_keys}{Columns that only apply to study events.}

\item{recall_keys}{Columns that only apply to recall events.}

\item{position_key}{Column indicating the position of each item in either
the study list or the recall sequence.}
}
\value{
Merged information about study and recall events. Each row
corresponds to one unique input/output pair.
}
\description{
Merge separated study and recall events that have the same subject, list,
and item.
}
\examples{
study <- data.frame(
  subject = c(1, 1), list = c(1, 1), position = c(1, 2), item = c("a", "b")
)
recall <- data.frame(subject = 1, list = 1, position = 1, item = "b")
merge_lists(study, recall)
}
