% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{reset_list}
\alias{reset_list}
\title{Reset list index}
\usage{
reset_list(data)
}
\arguments{
\item{data}{Raw or merged data. Must have subject and list fields.}
}
\value{
Data with a renumbered list field, starting from 1.
}
\description{
Reset list numbering to be sequential starting from one.
}
\examples{
# Data where the list number does not start at 1
subjects_list <- list(1, 1)
study_lists <- list(list("a", "b"), list("c", "d"))
recall_lists <- list(list("b"), list("c", "d"))
list_nos <- list(3, 4)
raw <- table_from_lists(
  subjects_list, study_lists, recall_lists,
  lists = list_nos
)
raw

# Reset the list number
reset_list(raw)
}
