% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{sample_data}
\alias{sample_data}
\title{Load a sample dataset}
\usage{
sample_data(study)
}
\arguments{
\item{study}{Study to load (options: "Morton2013").}
}
\value{
A data.frame with data from the specified \code{study}.
}
\description{
Load one of the included sample datasets.
}
\examples{
data <- sample_data("Morton2013")
head(data)
}
