% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{spc}
\alias{spc}
\title{Serial position curve}
\usage{
spc(data)
}
\arguments{
\item{data}{Merged study and recall data.}
}
\value{
Results with subject, input, and recall columns.
}
\description{
Recall probability as a function of serial position in the list.
}
\examples{
raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
recall <- spc(data)
head(recall)
}
