% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{split_lists}
\alias{split_lists}
\title{Convert data to split format.}
\usage{
split_lists(data, phase, keys = NULL, names = NULL, item_query = NULL)
}
\arguments{
\item{data}{Free recall data in raw or merged format.}

\item{phase}{Phase of free recall ('study' or 'recall') to split. If ‘raw’,
all trials will be included.}

\item{keys}{Data columns to include in the split data. If not specified, all
columns will be included.}

\item{names}{Name for each column in the returned split data. Default is to
use the same names as the input columns.}

\item{item_query}{Query string to select study trials to include.}
}
\value{
Data in split format. Each included column will be a key in the
named list.
}
\description{
Convert study, recall, or all events to list format.
}
\examples{
# Create raw and merged data
list_subject <- list(1, 1)
study <- list(list("absence", "hollow"), list("fountain", "piano"))
recall <- list(list("absence"), list("piano", "fountain"))
raw <- table_from_lists(list_subject, study, recall)
data <- merge_free_recall(raw)

# Get study events split by list, just including the list and item fields.
split_lists(data, "study", keys = list("list", "item"))

# Export recall events, split by list.
split_lists(data, "recall", keys = list("item"))

# Raw events (i.e., events that haven’t been scored) can also be exported to
# list format.
split_lists(raw, "raw", keys = list("position"))
}
