% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{table_from_lists}
\alias{table_from_lists}
\title{Create table format data from lists}
\usage{
table_from_lists(subjects, study, recall, lists = NULL, ...)
}
\arguments{
\item{subjects}{Subject identifier for each list.}

\item{study}{List of items for each study list.}

\item{recall}{List of recalled items for each study list.}

\item{lists}{List of list numbers. If not specified, lists for each subject
will be numbered sequentially starting from one.}

\item{...}{Additional arguments specify additional columns. Each must be a
list where the first item indicates study list values and the second item
indicates recall list values. If either item is NULL, that column will be
undefined for that phase.}
}
\value{
Data in table format.
}
\description{
Convert study and recall lists to table format.
}
\examples{
# Create standard columns from list data.
subjects_list <- list(1, 1, 2, 2)
study_lists <- list(
  list("a", "b"), list("c", "d"), list("e", "f"), list("g", "h")
)
recall_lists <- list(list("b"), list("d", "c"), list("f", "e"), list())
table_from_lists(subjects_list, study_lists, recall_lists)

# Include non-standard columns named col1 and col2.
subjects_list <- list(1, 1)
study_lists <- list(list("a", "b"), list("c", "d"))
recall_lists <- list(list("b"), list("d", "c"))
col1 <- list(list(list(1, 2), list(1, 2)), list(list(2), list(2, 1)))
col2 <- list(list(list(1, 1), list(2, 2)), NULL)
table_from_lists(
  subjects_list, study_lists, recall_lists,
  col1 = col1, col2 = col2
)
}
